refactor(core-transactions): merge v1 and v2 core transactions
