+++ packages/core-transactions/src/handlers/core/vote.ts
@@ -11,10 +11,11 @@ import {
 import { TransactionHandler, TransactionHandlerConstructor } from "../transaction";
 import { DelegateRegistrationTransactionHandler } from "./delegate-registration";
 
-// todo: revisit the implementation, container usage and arguments after core-database rework
-// todo: replace unnecessary function arguments with dependency injection to avoid passing around references
 @Container.injectable()
 export class VoteTransactionHandler extends TransactionHandler {
+    @Container.inject(Container.Identifiers.TransactionHistoryService)
+    private readonly transactionHistoryService!: Contracts.Shared.TransactionHistoryService;
+
     @Container.inject(Container.Identifiers.TransactionPoolQuery)
     private readonly poolQuery!: Contracts.TransactionPool.Query;
 
@@ -27,10 +28,48 @@ export class VoteTransactionHandler extends TransactionHandler {
     }
 
     public getConstructor(): Transactions.TransactionConstructor {
-        return Transactions.One.VoteTransaction;
+        return Transactions.Core.VoteTransaction;
     }
 
-    public async bootstrap(): Promise<void> {}
+    public async bootstrap(): Promise<void> {
+        const criteria = {
+            typeGroup: this.getConstructor().typeGroup,
+            type: this.getConstructor().type,
+        };
+
+        for await (const transaction of this.transactionHistoryService.streamByCriteria(criteria)) {
+            Utils.assert.defined<string>(transaction.senderPublicKey);
+            Utils.assert.defined<string[]>(transaction.asset?.votes);
+
+            const wallet = this.walletRepository.findByPublicKey(transaction.senderPublicKey);
+
+            for (const vote of transaction.asset.votes) {
+                const hasVoted: boolean = wallet.hasAttribute("vote");
+
+                let delegateVote: string = vote.slice(1);
+                if (delegateVote.length !== 66) {
+                    const delegateWallet: Contracts.State.Wallet = this.walletRepository.findByUsername(delegateVote);
+                    delegateVote = delegateWallet.getPublicKey()!;
+                }
+
+                if (vote.startsWith("+")) {
+                    if (hasVoted) {
+                        throw new AlreadyVotedError();
+                    }
+
+                    wallet.setAttribute("vote", delegateVote);
+                } else {
+                    if (!hasVoted) {
+                        throw new NoVoteError();
+                    } else if (wallet.getAttribute("vote") !== delegateVote) {
+                        throw new UnvoteMismatchError();
+                    }
+
+                    wallet.forgetAttribute("vote");
+                }
+            }
+        }
+    }
 
     public async isActivated(): Promise<boolean> {
         return true;
