+++ packages/core-transactions/src/handlers/index.ts
@@ -1,9 +1,8 @@
 import { TransactionHandlerProvider } from "./handler-provider";
 import { TransactionHandlerRegistry } from "./handler-registry";
-export * as One from "./one";
+export * as Core from "./core";
 export * as Solar from "./solar";
 import { TransactionHandler, TransactionHandlerConstructor } from "./transaction";
-export * as Two from "./two";
 
 export {
     TransactionHandler,
