--- packages/core-transactions/src/handlers/one/index.ts
@@ -1,4 +1,0 @@
-export * from "./transfer";
-export * from "./second-signature-registration";
-export * from "./delegate-registration";
-export * from "./vote";
