--- packages/core-transactions/src/handlers/two/index.ts
@@ -1,11 +1,0 @@
-export * from "./transfer";
-export * from "./second-signature-registration";
-export * from "./delegate-registration";
-export * from "./vote";
-export * from "./multi-signature-registration";
-export * from "./ipfs";
-export * from "./multi-payment";
-export * from "./delegate-resignation";
-export * from "./htlc-lock";
-export * from "./htlc-claim";
-export * from "./htlc-refund";
